+++ Libraries/LibGfx/DisjointRectSet.h
@@ -43,7 +43,7 @@ public:
     void add(const Rect&);
 
     bool is_empty() const { return m_rects.is_empty(); }
-    int size() const { return m_rects.size(); }
+    size_t size() const { return m_rects.size(); }
 
     void clear() { m_rects.clear(); }
     void clear_with_capacity() { m_rects.clear_with_capacity(); }
